refactor: csr 분리
