+++ src/app/landing/page.tsx
@@ -1,5 +1,14 @@
+import LandingTracking from '@/views/landing/LandingTracking';
+import StartButton from '@/views/landing/StartButton';
 import LandingView from '@/views/LandingView';
 
 export default async function Home() {
-  return <LandingView />;
+  return (
+    <>
+      <LandingView>
+        <StartButton />
+      </LandingView>
+      <LandingTracking />
+    </>
+  );
 }
